# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Sudoku Game API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /profiles:
    post:
      operationId: createProfile
      tags: [profiles]
      summary: Create a user profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProfileInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Profile"
  /profiles/sync:
    post:
      operationId: syncProfile
      tags: [profiles]
      summary: Find or create a profile by Clerk user ID or device ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProfileSync"
      responses:
        "200":
          description: Profile found or created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Profile"
  /profiles/{id}:
    get:
      operationId: getProfile
      tags: [profiles]
      summary: Get a user profile
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Profile"
        "404":
          description: Not found
    patch:
      operationId: updateProfile
      tags: [profiles]
      summary: Update a user profile
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProfileUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Profile"
...
  /profiles/search:
    get:
      operationId: searchProfiles
      tags: [profiles]
      summary: Search profiles by username
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
        - name: exclude
          in: query
          required: false
          schema:
            type: integer
          description: Profile ID to exclude from results
      responses:
        "200":
          description: Matching profiles
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProfileSummary"
components:
  schemas:
...
    Profile:
      type: object
      required: [id, username, theme, createdAt]
      properties:
        id:
          type: integer
        username:
          type: string
        avatar:
          type: ["string", "null"]
        theme:
          type: string
          enum: [light, dark]
        highlightErrors:
          type: boolean
        showTimer:
          type: boolean
        gems:
          type: integer
          description: "Accumulated gem balance usable for future in-game features"
        createdAt:
          type: string
    ProfileInput:
      type: object
      required: [username]
      properties:
        username:
          type: string
          minLength: 1
          maxLength: 30
        avatar:
          type: string
        theme:
          type: string
          enum: [light, dark]
        highlightErrors:
          type: boolean
        showTimer:
          type: boolean
    ProfileUpdate:
      type: object
      properties:
        username:
          type: string
          minLength: 1
          maxLength: 30
        avatar:
          type: string
        theme:
          type: string
          enum: [light, dark]
        highlightErrors:
          type: boolean
        showTimer:
          type: boolean
    ProfileSync:
      type: object
      properties:
        clerkUserId:
          type: string
        deviceId:
          type: string
        username:
          type: string
        avatar:
          type: string
...
    ProfileSummary:
      type: object
      required: [id, username]
      properties:
        id:
          type: integer
        username:
          type: string
        avatar:
          type: ["string", "null"]
        gems:
          type: integer
...