# lib/api-zod/src/generated/types/profileSummary.ts
/**
 * Generated by orval v8.15.0 🍺
 * Do not edit manually.
 * Api
 * Sudoku Game API
 * OpenAPI spec version: 0.1.0
 */

export interface ProfileSummary {
  id: number;
  username: string;
  /** @nullable */
  avatar?: string | null;
  gems?: number;
}